Create Ringr Lambda functions
Creates the initial AWS Lambda functions for Ringr, including functions
to add and remove endpoints, a function to manage endpoints by
responding to DynamoDB streams, and a function for publishing a
notification to the Ringr SNS topic. These functions should provide the
basic functionality for Ringr.

Future development should provide configuration-as-code in the form of
CloudFormation or Terraform to define the necessary IAM, API Gateway,
DynamoDB, and SNS components for utilizing these functions. It should
further focus on adding tests to ensure these and any additional
functions operate correctly.